CMake Setup, Problems 15,16,17

+++ Problem14/Problem14.h
@@ -2,5 +2,5 @@
 
 class Problem14 {
 public:
-	static int Solve(int N);
+	static long long Solve(long long N);
 };
